Upstream changes (#12)
* Updated How to pay content

* Count of orgs and live services for platform admin

This makes it consistent that an option which contains more options has
a hint about how many options it contains.

Also adds a formatter to get us ready for 1,000 services 🎉

* Make one method for comma-formatting numbers

We were doing this a few different ways in different places.

* Update live services count when service is counted

If we change our mind and decide whether a service should/should not be
counted in the list of live services then we should also drop the cache
which stores the count of how many live services there are.

* Update app/templates/views/pricing/how-to-pay.html

Co-Authored-By: Chris Hill-Scott <me@quis.cc>

* Update app/templates/views/pricing/how-to-pay.html

Co-Authored-By: Chris Hill-Scott <me@quis.cc>

* Update app/templates/views/pricing/how-to-pay.html

Co-Authored-By: Chris Hill-Scott <me@quis.cc>

* Add abbreviations

* Stop `JSONModel` hiding attribute errors

`__getattr__` is called whenever an attribute error is raised.

This means that if something deep inside a property on a model raised
an attribute error, that error would be caught by `__getattr__`, which
would then raise an exception that looked like the property itself
didn’t exist. Very confusing.

The solution seems to be to override `__getattribute__` instead, which
handles _all_ attributes, not just those that aren’t explicitly defined.
We then only intervene if the desired attribute is one of the
`ALLOWED_PROPERTIES`, otherwise falling through to the built in methods
of the underlying `object`.

* Change 'email reply-to' to 'reply-to email'

* Update content

Update references to 'reply-to email' and remove negative contractions

* Update 'reply-to email address'


* Fix tests

* Update purchase order content

* Add link to pricing

* Update 'reply-to email address'

* Add contact us

* Update 'contact us'

* Update 'contact us' content

* Fix tests

* Make ‘thanks’ page consistent with support page

On the support page we now promise that we’ll:
- look at tickets within half an hour
- reply within one working day

The thanks page was always promising a reply within half an hour, during
business hours or an emergency. This commit changes the ‘thanks’ page to
be consistent with the support page.

* Update 'your message'