Sync upstream (#19)
* Updated How to pay content

* Go back to template if coming from template

Most users don’t have multiple contact blocks. So by default it should
feel like you’re just editing the one contact block, rather than
managing a collection of them. So this page skips the ‘choose’ page when
the user doesn’t yet have any contact blocks.

* Go straight to add if no contact blocks yet

Same as previous commit, but for going forward, not back.

* Always offer option to add new sender

Not sure why we had this logic. We shouldn’t make you jump over to the
settings page if none of the available options are right for this
letter.

* Apply sender to template when adding

If you’ve come from a template to add a new letter sender then it’s
because you want those words on that template. This commit adds the
extra API call to make that happen.

* Remove unused variable

Presume this is from when we used to render the letter preview in a
different way. Removing this saves us an API call every time we’re
loading this page.

* Refactor back link into view layer

* Count of orgs and live services for platform admin

This makes it consistent that an option which contains more options has
a hint about how many options it contains.

Also adds a formatter to get us ready for 1,000 services 🎉

* Make one method for comma-formatting numbers

We were doing this a few different ways in different places.

* Update live services count when service is counted

If we change our mind and decide whether a service should/should not be
counted in the list of live services then we should also drop the cache
which stores the count of how many live services there are.

* Update app/templates/views/pricing/how-to-pay.html

Co-Authored-By: Chris Hill-Scott <me@quis.cc>

* Update app/templates/views/pricing/how-to-pay.html

Co-Authored-By: Chris Hill-Scott <me@quis.cc>

* Update app/templates/views/pricing/how-to-pay.html

Co-Authored-By: Chris Hill-Scott <me@quis.cc>

* Add abbreviations

* Stop `JSONModel` hiding attribute errors

`__getattr__` is called whenever an attribute error is raised.

This means that if something deep inside a property on a model raised
an attribute error, that error would be caught by `__getattr__`, which
would then raise an exception that looked like the property itself
didn’t exist. Very confusing.

The solution seems to be to override `__getattribute__` instead, which
handles _all_ attributes, not just those that aren’t explicitly defined.
We then only intervene if the desired attribute is one of the
`ALLOWED_PROPERTIES`, otherwise falling through to the built in methods
of the underlying `object`.

* Change 'email reply-to' to 'reply-to email'

* Update content

Update references to 'reply-to email' and remove negative contractions

* Update 'reply-to email address'

* Update 'reply-to email address'

* Updated 'reply-to email address'

* Update 'reply-to email address'

* Update 'reply-to email address'

* Fix tests

* Update purchase order content

* Add link to pricing

* Update 'reply-to email address'

* Add contact us

* Update 'contact us'

* Update 'Contact us'

* Update 'contact us'

* Update 'contact us'

* Update 'contact us' content

* Update 'contact us' content

* Update 'contact us' content

* Update 'contact us' content

* Fix tests

* Make ‘thanks’ page consistent with support page

On the support page we now promise that we’ll:
- look at tickets within half an hour
- reply within one working day

The thanks page was always promising a reply within half an hour, during
business hours or an emergency. This commit changes the ‘thanks’ page to
be consistent with the support page.

* Update 'your message'

* Add a tiny bit more spacing under the folder heading when the user has no template search or channel filters

* Make query string comparison ignore order

Query string ordering is non-deterministic. This can cause tests to fail
in a non helpful way when we’re comparing two URLs. The values in the
query string can match, but the string won’t because the order is
different. This commit adds some code to split up the URL and check that
each part of it matches, rather than checking the thing as a whole.

* Stop allowing the service org type to be changed

The service organisation type will either be the same as the org type of
the service's organisation or will be set by a user when creating a new
service. This removes the ability to change it from the platform admin
settings table.

* Add organisation_type property to Service model

This will return the organisation_type of the service's organisation (if
there is one), or the organisation_type of the service if not.

* skip test as -> not using zendesk