FIX: Revert h2 to legend
Following the dictates of issue 56 [1], this commit refactors the backend
and frontend rendering of the h2 title, folder summary, hidden legend and
folder permission checkboxes.

The rendering and DOM layout previously was structure like this
+---form
	| ---h2
	| ---div.selection-summary
	| ---div.form-group
		| +---fieldset
			| +---legend
			| +---div.selection-nested
				| +---ul

This aforementioned is confusing as it was highlighted as a label, but is
visually heading, this isn't fit for users that use screen readers and other
aids.

This commit makes changes to the backend and frontend to balance
percieved constraints and removed identified redundant code that is
now structured as follows:

+---form
	| ---div.form-group
		| +---fieldset
			| +---legend
			| ---div.selection-summary
			| +---div.selection-nested
				| +---ul

[1] https://github.com/cds-snc/notification-planning/issues/56