Integrate Mixpanel within Notify Admin backend
Given the need to track users activities on Notify Admin, we intend to
integrate Mixpanel Python library that would asist with tracking certain
user events within the Notify Admin.

This commit focuses on integrating with the Notify admin backend and not with
the front-end mainly due to blocking nature of Javascript code.