External links are now boolean values rather than passing in an image
External links have a little icon beside them and so far we have been
passing in image paths when a link points somewhere external.

This isn't really needed: the image path could easily be hard-coded in
the macro itself and that makes the macro easier to use.

It also makes it easier to build menus from pure data, incidentally.