Add 2 new image domains: canada.ca and cdssandbox.xyz (#1216)
We are planning to receive content from the GC articles API, including
images. However, currently images can only come from "self" or the
asset domain, so we have been linking to images already hosted on Notify
from GC Articles.

However, this causes problems moving between environments.
For example, if we are linking to an image hosted on "notification.cdssandbox.xyz",
then it won't show up locally, on Heroku, or in Production.