Refactor getting a preview URL, remove "g.preview"
We use the preview URL for the "Edit this page" link in the preview
banner.

Previously, I was setting 2 variables: one was a boolean set to True if
we were in preview mode, and the second was the preview URL.

It occurred to me that we would never have a preview URL if preview was
false, so we can just check for the existence of a preview URL as a
truthy value and delete the other param.

The other change here was to move the URL generation into its own function
in "app.articles". It needs to know what the current language is so that
WordPress can generate the right Preview link, and all the other functions
in app.articles are getting the current language themselves rather than
having the view do anything with it.

Refactor the get preview URL method