Explicitly 404 if there is no translated page_id
If there is no translated page available, get_lang_url would default
to returning the current page slug or the current ID.

If it returns the current page slug, WordPress will send back a 404
response (because it doesn't see that slug for the other language),
but the ID would just return as normal.

Speaking to Elise, we preferred to clearly indicate something is
wrong rather than make it look like a page exists but wasn't translated yet.