fix: remove wait for function active
The wait for function active call is timing out on old
functions, so it's being removed.

This was originally added to handle a case where this workflow
was called on a function that was just created, but since
the workflow runs in the middle of the night, the risk
of this scenario is low.