fix: create PR review env log group
Update the PR review environment deploy to also create the
CloudWatch log group and set a retention period.

This is because by default, the log group is only created when
the PR review env function is invoked which was leading to failures
cleaning up old review envs trying to delete non-existent
log groups.