fix: build PR review Docker image with `docker build` (#1471)
The Docker images being built by the GitHub Docker build and push
action are no longer working with AWS Lambda.  We get the following
error when attempting to create a function with them:

```
The image manifest or layer media type for the source image
$IMAGE_TAG is not supported.
```

Using a raw `docker build` command works as expected and is
consistent with how all the other Notify Docker images are built.