Fix bug where the report url expires early (#2174)
Fix a bug where the report url expires early.

We have been using the presigned s3 url that is saved in the db. But it is not actually necessary since the user now no longer downloads the report directly from s3. They download it via the download_report_csv endpoint in admin so we can set the appropriate filename. Admin already has access to this bucket, so we can just get the s3 object directly.