Fix: SMS daily limit check for CSV sending should not use fragments (#2180)
This commit changes the validation we do when a user uploads a CSV for an sms send. We now just look at the number of rows using the len function defined here: https://github.com/cds-snc/notification-utils/blob/main/notifications_utils/recipients.py#L114-L117

The daily_sms_count in redis is already being correctly set to the number of sms sent, not to the number of sms fragments.