Get annual email/sms totals from redis if available (#2196)
The get_monthly_notification_stats api call is taking ~1 second for some services. This data is monthly, but we are just aggregating in annually, adding today's totals, and using it to display SMS/email sent this year on the dashboard. We already have this data in Redis, if the service has sent today. So we can use that if it is available, and greatly reduce the load time for the dashboard.

This change:
- adds a get_annual_data function, that gets data from redis if available, else it gets it from the same api call as before
- adds unit tests for get_annual_data
- adds new type definitions to keep track of all the data structures