Feat/newsletter subscribe (#2463)
This PR implements the Figma design for the new "subscribe to our newsletter" form element.

This change will be hidden in prod because of the {% if isHome and config["NOTIFY_ENVIRONMENT"].lower() != "production" %} check.

Some of the functionality is still to do. I will make a future PR when the api endpoints become available.

Working on cds-snc/notification-planning#2624